chore: settings refactor
